fix: Extract fieldRelations from final schema
Previously, fieldRelations were extracted during transformation via
distilation hooks. This was fast, as there were no need to iterate over
the types more that once, but it produces wrong data for
interesections, where I merged the fields _after_ the transformation,
and generated data therefor contained only the parts of the
intersection which don't exists in the final schema and did not
contained the required merged type.

I could merge the schema before the transformation, but it proved to
have many difficulties. As I do not need super fast runtime and the
distillation is difficult even without performance optimizations, I
decided to migrate to pattern, where the distillation will only generate
GQL Schema (with all data needed for our gqlEngine), and the data
needed for GQL runtime will be later ingested from the Schema via
introspection. This will nicely separate concerns and will greatly
reduce the complexity of distillation.